Data Validation

+++ Controllers/StockController.cs
@@ -13,11 +13,11 @@ namespace api.Controllers
 {
     [Route("api/stock")]
     [ApiController]
-    public class StockController:ControllerBase
+    public class StockController : ControllerBase
     {
         readonly ApplicationDBContext _context;
         private readonly IStockRepository _stockRepo;
-        public StockController(ApplicationDBContext context,IStockRepository stockRepo)
+        public StockController(ApplicationDBContext context, IStockRepository stockRepo)
         {
             _stockRepo = stockRepo;
             _context = context;
@@ -26,35 +26,43 @@ namespace api.Controllers
         [HttpGet]
         public async Task<IActionResult> GetAll()
         {
+            if (!ModelState.IsValid)
+                return BadRequest(ModelState);
             var stocks = await _stockRepo.GetAllAsync();
-            var stockDto= stocks.Select(s=>s.ToStockDto());
+            var stockDto = stocks.Select(s => s.ToStockDto());
             return Ok(stockDto);
         }
 
-        [HttpGet("{id}")]
-        public async Task<IActionResult> GetById([FromRoute]int id)
+        [HttpGet("{id:int}")]
+        public async Task<IActionResult> GetById([FromRoute] int id)
         {
+            if (!ModelState.IsValid)
+                return BadRequest(ModelState);
             var stock = await _stockRepo.GetByIdAsync(id);
-            if(stock == null) return NotFound();
+            if (stock == null) return NotFound();
             return Ok(stock.ToStockDto());
         }
 
         [HttpPost]
         public async Task<IActionResult> Create([FromBody] CreateStockRequestDto stockDto)
         {
-            var stockModel= stockDto.ToStockFromCreateDTO();
-           await _context.Stocks.AddAsync(stockModel);
-           await _context.SaveChangesAsync();
+            if (!ModelState.IsValid)
+                return BadRequest(ModelState);
+            var stockModel = stockDto.ToStockFromCreateDTO();
+            await _context.Stocks.AddAsync(stockModel);
+            await _context.SaveChangesAsync();
 
-            return CreatedAtAction(nameof(GetById),new {id=stockModel.Id},stockModel.ToStockDto());
+            return CreatedAtAction(nameof(GetById), new { id = stockModel.Id }, stockModel.ToStockDto());
         }
 
         [HttpPut]
-        [Route("{id}")]
+        [Route("{id:int}")]
         public async Task<IActionResult> Update([FromRoute] int id, [FromBody] UpdateStockRequestDto updateDto)
         {
+            if (!ModelState.IsValid)
+                return BadRequest(ModelState);
             var stockModel = await _context.Stocks.FindAsync(id);
-            if(stockModel == null) return NotFound();
+            if (stockModel == null) return NotFound();
 
             stockModel.Symbol = updateDto.Symbol;
             stockModel.CompanyName = updateDto.CompanyName;
@@ -68,11 +76,13 @@ namespace api.Controllers
         }
 
         [HttpDelete]
-        [Route("{id}")]
+        [Route("{id:int}")]
         public async Task<IActionResult> Delete([FromRoute] int id)
         {
+            if (!ModelState.IsValid)
+                return BadRequest(ModelState);
             var stockModel = await _context.Stocks.FindAsync(id);
-            if(stockModel == null) return NotFound();
+            if (stockModel == null) return NotFound();
 
             _context.Stocks.Remove(stockModel);
             await _context.SaveChangesAsync();
